seperated import command into keys and mnemonics
